+++ packages/application/src/plugin/queries/get-plugin/get-plugin.result.ts
@@ -4,6 +4,7 @@ export interface GetPluginResult {
   readonly pluginId: string;
   readonly name: string;
   readonly version: string;
+  readonly icon: string;
   readonly state: PluginRuntimeState;
   readonly enabled: boolean;
 }